Skip freebins if they are more wasteful than free shelves (closes #25)
This also refactors the packOne() code a little bit to break out
`allocFreebin()` and `allocShelf()` into separate functions.

The new code exits a little faster when the requested bin width/height
matches up perfectly with a free bin or free shelf, which can improve perf
slightly (so I also updated the benchmark numbers in the README).